+++ src/components/LogoutPage.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { useDispatch } from "react-redux";
-import { logout } from "../../store/user/actions";
+import { logout } from "../store/user/actions";
 
 export default function Logout() {
   const dispatch = useDispatch();
