+++ src/components/Map/MapContainer.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import { useSelector } from "react-redux";
 import { GoogleMap, withGoogleMap } from "react-google-maps";
 // import { usePosition } from "./UserLocation";
